Improve development workflow
Add a GameFolder variable to be used to define where to find the
Bannerlord installation. This way it is easier for other developers to
contribute to the project. Another step would be to get rid of having to
hard code this path within the project file all together.

By using this above approach within the .csproj.user file it is also
possible to setup the development environment in a way that it can
launch the Bannerlord configured to run the Mod right out of the IDE.
This change makes the debuglink script obsolete. See README for more
information about how to do this.

Also add a PostBuildEvent that takes the build artifacts and moves them
so the final product is in the required structure for a release. This
change allows us to simplify the release script.

- Change msbuild to msbuild.exe within release script
- Rename assembly to the full name to increase verbosity
- Fix a typo in Strings file
- Remove a duplicated assembly reference